pyproject.toml
[tool.poetry]
name = "sample-retail"
version = "0.1.0"
description = ""
authors = ["Craig Lenzen <lenzenc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
